Created comments
- Created comments
- Moved CheckUser to utils
