+++ src/components/PaginationTasks/index.js
@@ -6,6 +6,8 @@ import {
   BsChevronLeft,
   BsChevronRight,
 } from 'react-icons/bs';
+
+// Recoil
 import { useRecoilState } from 'recoil';
 import { currentPageState } from '../../recoil/todos';
