+++ src/containers/Home/index.js
@@ -1,8 +1,14 @@
 import React from 'react';
 import { Flex } from 'theme-ui';
+
+// Containers
 import NewTaskPanel from '../NewTaskPanel';
 import TasksListPanel from '../TasksListPanel';
+
+// Skeletons
 import SkeletonTasksListPanel from '../../components/SkeletonTasksListPanel';
+
+// ErrorBoundarys
 import ErrorBoundaryForList from '../../components/ErrorBoundaryForList';
 
 function Home() {
