+++ src/recoil/user.js
@@ -1,5 +1,5 @@
 import { atom, selector } from 'recoil';
-import { handleCheckUser } from '../api/user';
+import { handleCheckUser } from '../utils';
 import { DEFAULT_USER } from '../consts';
 
 export const checkUserQuery = selector({
